+++ R/ggpreview.R
@@ -2,7 +2,7 @@
 #'
 #' This function previews a ggplot in its actual dimensions in order to see how
 #' it will look when saved. It is also significantly faster than the default
-#' preview in RStudio for ggplots created using cfbplotR.
+#' preview in RStudio for ggplots created using cfbplotR. This is copied directly from `nflplotR`
 #'
 #' @inheritParams ggplot2::ggsave
 #' @param asp The aspect ratio of the plot calculated as `width / height`. If